LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/stretch/flex-line-005-expected.txt


FAIL .stretch 1 assert_equals:
<div class="container">
  <div class="stretch" style="width: 75px; border-color: cyan" data-expected-height="156"></div>
  <div style="border-color: orange; height: 150px"></div>
</div>
height expected 156 but got 578
